completed chapter 4
