+++ index.py
@@ -1,0 +1,3 @@
+from slithering import Frog, Goldfish, Goose, Mallard, Swan
+from swimming import Copperhead, JellyFish, PufferFish, RatSnakes, Swordfish
+from walking import Alpaca, Camel, Donkey, Goat, Llama
