+++ slithering/__init__.py
@@ -1,0 +1,5 @@
+from .frog import Frog
+from .goldfish import Goldfish
+from .goose import Goose
+from .mallard import Mallard
+from .swan import Swan
