+++ swimming/__init__.py
@@ -1,0 +1,5 @@
+from .copperhead import Copperhead
+from .jellyfish import JellyFish
+from .pufferfish import PufferFish
+from .ratsnakes import RatSnakes
+from .swordfish import Swordfish
