+++ swimming/copperhead.py
@@ -1,0 +1,9 @@
+from datetime import date
+
+class Copperhead:
+
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today
+        self.swimming = True
