+++ swimming/jellyfish.py
@@ -1,0 +1,10 @@
+from datetime import date
+
+
+class JellyFish:
+
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today
+        self.swimming = True
