+++ swimming/pufferfish.py
@@ -1,0 +1,10 @@
+# import the python datetime module to help us create a timestamp
+from datetime import date
+
+class PufferFish:
+
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today
+        self.swimming = True
