+++ walking/__init__.py
@@ -1,0 +1,5 @@
+from .alpaca import Alpaca
+from .camel import Camel
+from .donkey import Donkey
+from goat import Goat
+from llama import Llama
